pyproject.toml
[tool.poetry]
name = "django-user-service"
version = "0.1.0"
description = "A Django user service."
authors = ["Anthony Honstain"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
django-structlog = "^7.1.0"

[tool.poetry.dev-dependencies]
black = "^24.2.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 120